requirements_base.txt
otree-core[mturk]=1.1.2
Django==1.8.8 # for heroku, needs to be explicitly in requirements file
numpy >= 1.8.0
scipy >= 0.19.1
matplotlib >= 1.3.1